pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "thermal-tracker"
version = "0.1.0"
description = "Virtual Multi-View Thermal Tracking Simulation"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "opencv-python-headless>=4.8",
    "scipy>=1.11",
    "scikit-learn>=1.3",
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
